chore: move package-manager.ts inside global utils
